requirements.txt
aiogram==3.13.1
sqlalchemy==2.0.30
aiosqlite==0.19.0
python-dotenv==1.0.0